python/requirements.txt
matplotlib==3.7.1
perlin_noise==1.12
Pillow==9.5.0
Pillow==9.5.0
